[SPEC/CODE_SMELL] Stubbed out Warden sessions
[SPEC] Stubbed out Warden::SessionSerializer
[SPEC] Stubbed out Rails::Sessions::AbstractStore

This was necessary to run the test. However, this is not the right thing
to do. The current code uses its own serialization from Redis to pull
the account.id. This forces users of Intermodal to use a redis store.

Properly, we serialize using Warden, which allows Intermodal users to
drop in their own store. This does mean changing the X-Auth-Token
strategy to work with this.

For now, in the interest of getting specs and engines fully working, I'm
leaving this here and adding the issue to Github.