Intermodal::API now an Engine-like Railtie instead of a mixin for an Engine
Rails 3.0 was ambiguous about what an Engine is. Rails 4.0 Engines are
firmly implemented as a miniature Rails-within-Rails. It assumes
that Engines have app paths, classes that are loaded, and whatnot.

Rather than doing that, we make Intermodal::API a railtie that pulls in
enough code from Rails::Engine to act as an endpoint without the
assumption that the API will have paths.

Although this lays the groundwork for multiple versions of APIs, it
isn't complete without isolate_namepace. Rather than implementing that
now, we'll keep it like this until an app that requires multiple
versions require it. Cross that bridge when we come to.

You also no longer need to call routes.append. You can directly
describe routes in the engine with:

  routes do
    resources :items
  end

Each API will also do it's own loading for controllers.

Other things that can be improved including implementing eager_load!
by dumping the dynamically created classes and re-evaluating the API
file itself.